Revert "Solves #3: Vans filter doesn't work"
