+++ src/App.css
@@ -79,7 +79,7 @@ h4 {
   align-items: center;
   justify-content: space-between;
   color: var(--font-color-dark);
-  width: 160px;
+  width: 130px;
   font-size: .9rem;
   font-weight: 400;
   margin: 2em 0;
@@ -186,7 +186,6 @@ h4 {
 }
 
 /* ========== LAYOUT ABOUT ========== */
-
 .about-container {
   background-size: cover;
 }
@@ -312,28 +311,6 @@ h4 {
   margin: 0;
 }
 
-.van-info-section {
-  display: flex;
-  flex-direction: column;
-}
-
-@media (min-width: 700px) {
-  .van-info-section img {
-    max-width: 30%;
-  }
-
-  .van-info-section {
-    flex-direction: row;
-  }
-
-  .van-info-section-content {
-    padding: 2em;
-  }
-
-  .van-info-section-content button {
-    max-width: 500px;
-  }
-}
 
 /* ========== LAYOUT HOST PAGE ========== */
 .host-layout-container {
@@ -428,7 +405,6 @@ h4 {
   text-decoration: underline;
 }
 
-
  /* GALLERY */
 
 .host-vans-gallery {
@@ -438,16 +414,6 @@ h4 {
 }
 
 /* REVIEWS */
-
-/* ========== LAYOUT NOTFOUND ========== */
-.notfound-page {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  /* margin: auto; */
-  padding: 2em;
-  max-width: 700px;
-}
 
 /* ========== LAYOUT FOOTER ========== */
 footer {
